# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ShareWallet Payments API
  version: 1.0.0
  description: |
    All endpoints require a Supabase session cookie (set by `/auth/callback`).
    Every ledger write runs inside a Postgres function, so each call is atomic.
servers:
  - url: /api/v1
components:
  securitySchemes:
    supabaseSession:
      type: apiKey
      in: cookie
      name: sb-<project-ref>-auth-token
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        issues: { type: array, items: { type: object } }
      required: [error]
...
  responses:
    Unauthorized:
      description: No signed-in user
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Invalid:
      description: Request body failed validation
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
security:
  - supabaseSession: []
paths:
...
  /wallets:
    post:
      summary: Create a wallet (caller becomes admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 80 }
                currency: { type: string, minLength: 3, maxLength: 3, default: USD }
                approvalThreshold: { type: integer, minimum: 1, maximum: 20, default: 2 }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  walletId: { type: string, format: uuid }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/Invalid' }
...